fix eslint and update files structure
