--- src/components/cardDetails.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-
-export default class carDetails extends Component {
-  render() {
-    return (
-      <div>{this.props.title}</div>
-    )
-  }
-}
